requirements.txt
discord.py==1.7.2
requests==2.26.0
